Implemented new TEI renderer
This cleans up the rendering within cts.Passage. In addition, I
have added <t o="..."> for the token's offset in the passage (sans
whitespace). This is not immediately useful but will be used for
quicker indexing internally.

Finally, I have slightly modified how we keep token indexes to
better adhere to the CTS spec about substrings. Each token is chopped
up into all of its substrings and indexed this way.

Refs #162