use border to prevent text shifting
Minor CSS change to use a border instead of padding in order to prevent
text shifting around. Before this change, a selected word took up
slightly more space than an unselected word. This caused to the right of
a word to shift every time a word was selected. The pixel difference is
small, but it's pretty noticeable as a user when selecting different
words.